Updated readme and reworked example for v5.3
