--- example/.storybook/addons.ts
@@ -1,1 +1,0 @@
-import "storybook-addon-styled-component-theme/dist/register";
