--- example/.storybook/config.ts
@@ -1,31 +1,0 @@
-import {addDecorator, configure} from "@storybook/react";
-import {withThemesProvider} from "storybook-addon-styled-component-theme";
-
-const defaultTheme = {
-  name: "DEFAULT",
-  backgroundColor: "azure",
-  textColor: "dimgrey",
-  borderRadius: "30px",
-};
-
-const darkTheme = {
-  name: "DARK",
-  backgroundColor: "black",
-  textColor: "seashell",
-  borderRadius: "100px",
-};
-
-export const getAllThemes = () => {
-  return [defaultTheme, darkTheme];
-};
-
-addDecorator(withThemesProvider(getAllThemes()));
-
-// automatically import all files ending in *.stories.js
-// @ts-ignore
-const req = require.context("../stories", true, /\.stories\.tsx?$/);
-function loadStories() {
-  req.keys().forEach((filename) => req(filename));
-}
-
-configure(loadStories, module);
