+++ example/.storybook/main.js
@@ -1,0 +1,4 @@
+module.exports = {
+    stories:['../stories/**/*.stories.(tsx|mdx)'],
+    addons: ['@storybook/preset-create-react-app','storybook-addon-styled-component-theme/dist/register'],
+  };
