+++ example/package.json
@@ -10,9 +10,13 @@
   "license": "ISC",
   "devDependencies": {
     "@babel/core": "^7.7.5",
-    "@storybook/addons": "^5.2.8",
-    "@storybook/react": "^5.2.8",
+    "@storybook/addons": "^5.3.13",
+    "@storybook/preset-create-react-app": "^1.5.2",
+    "@storybook/react": "^5.3.13",
     "babel-loader": "^8.0.6",
-    "storybook-addon-styled-component-theme": "link:../"
+    "storybook-addon-styled-component-theme": "1.3.0",
+    "react-scripts": "^3.4.0",
+    "styled-components": "^5.0.1",
+    "typescript": "^3.7.5"
   }
 }
